backend/requirements.txt
Django>=4.2,<5.1
wagtail>=6.0,<6.1
python-decouple==3.8
boto3==1.34.67
django-storages==1.14.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1